clean before building in generate_grammar_kit.py
Summary:
The grammar generation in `generate_grammar_kit.py` was not cleaning
the previous output before writing new output, so when the grammar
changed, old files would linger in the output directory.

Also replaced relative paths with absolute paths to make it possible
to copy/paste the commands from the script's output.

Reviewed By: asp2insp

fbshipit-source-id: 85bbcf8ddb