Fix compilation-database woes for cpp_python_extension targets.
Summary:
In the course of testing howtoeven linters, I noticed that frequently
linter fail to process many C++ files in an enlistment.
After some false starts, my debugging suggested that part of the problem
related to directories that contained cpp_python_extension targets. If
you try to build a `#compilation-database` rule for one of these targets
then Buck dies with an error about a missing-but-required Python Platform
argument for the target.

So this is an attempt to be minimally invasive and provide a "sane"
default for the Python Platform (`py-default`) to support generation
of a compilation database for these targets.

I let their C++ platform default to whatever Buck calculates, and I just
add a default Python platform for now.

What I would like to do as a separate work item is consider how to change
the compilation-database implementation in Buck so that it picks up whatever
the defaults would be for both C/C++ platform and Python Platform (as per the
codegen build) and uses those. Currently, the script invoked by linter/howtoeven
passes some `-c` arguments to Buck that forces everyone to use the same C++
platform flavor. I suspect this is causing lint-build failures for some projects.

However, from looking at the way flavors get
picked up and assigned, I suspect that work item will be a little trickier.

Reviewed By: styurin

fbshipit-source-id: 9beadb2da1