Don't print compiled filenames on Windows
Summary: When building C++ project with Buck on Windows cl.exe prints compiled filenames in logs. This means that for each file that is not cached there is an unnecessary line with the name of the file (not even the path). That produces too much output, specially in big projects, which can be hidden.

Reviewed By: styurin

fbshipit-source-id: b3c2eb81f3