Add ToolchainTypeInfo
Summary:
This basically treats the toolchain as just an opaque name for
serialization/deserialization. This is a little bit non-ideal, but it's the
best that we can do for now.

As we improve toolchains, we should be able to change toolchain
references in rules to actually be references to the underlying values/tools
and then this form of serialization becomes unnecessary.

Reviewed By: styurin

fbshipit-source-id: 173c8a88dd