Make JavacFactory and JavaOptions support parse time deps
Summary:
This just adds the required functions to these things and then properly
calls them in all the descriptions that use them. This is required to
allow these things to depend on build targets (i.e. to have a
python_binary implement javac).

Sidenote: This is pretty terrible and we need to figure out a way to
make it less manual/error-prone.

Reviewed By: ttsugriy

fbshipit-source-id: a40e320bed