Make ExternalJavacFactory support a BinaryBuildRule javac
Summary:
When a tool path is a BuildTargetSourcePath, we should be treating it as
a reference to a BinaryBuildRule and get the Tool via
BinaryBuildRule.getExecutableCommand.

This makes ExternalJavacFactory do that (and renames it to something
more appropriate).

Currently, only JavacSpecs derived from target args can have a
BuildTargetSourcePath for the configured Javac, and so this only affects
them. A followup will add that support to the specs derived from
.buckconfig.

Reviewed By: styurin

fbshipit-source-id: 035f6fb2fc