Make OutputsMaterializer support duplicate output directories
Summary:
If the returned Tree contains duplicate Directory nodes (which maybe
shouldn't be allowed), the materialization currently crashes when
creating the child map (map of digest -> Directory node) because we try
to insert multiple entries. As we broadly assume that multiple nodes
with the same digest are equivalent, it's safe to just accept duplicate
keys. Update the materializer to allow that.

Reviewed By: ttsugriy

fbshipit-source-id: c3c9bc8cb0