Make ProjectWorkspace handle exceptions consistently
Summary:
For most exceptions, ProjectWorkspace will catch the exception, print
the output to stderr and return an exit code just like actually calling
buck would do. But, oddly, for some exceptions it doesn't do that and
just propagates the exceptions. This requires users to detect errors in
two different ways and to either just fiddle around until they get it
right, or somehow memorize exactly the behavior of ProjectWorkspace.
Either way, they end up with really inconsistent looking tests.

This also has the downside that wrapping an exception in a
BuckUncheckedExecutionException (or any of the similar WrapsException
exceptions) can greatly change the behavior of ProjectWorkspace (despite
the fact that Buck itself handles it correctly).

This changes ProjectWorkspace to catch all exceptions that are
propagated out of Main and print them (with an ErrorLogger) to stderr
and return an appropriate ExitCode. We should, at some point, refactor
Main so that the error handling is actually shared between these two.

This updates all the tests to consistently get a ProcessResult and check
its ExitCode and Stderr.

Reviewed By: ttsugriy

fbshipit-source-id: d67322277d