handle prebuilt frameworks when outputting linker flags
Summary:
Turns out some targets are including a buck target for a prebuilt framework in the `frameworks` field which with the changes for BUCK_LINKER_FLAGS_SYSTEM will cause buck project to fail so we need to resolve the targets also.

Also when `--config apple.link_libraries_as_flags=true` then don't duplicate the fwks to the linker phase as they will be added  to OTHER_LDFLAGS via BUCK_LINKER_FLAGS_SYSTEM and make sure the system fwks/libs appear first in the linker list.

Example:

    frameworks = [
        "$SDKROOT/System/Library/Frameworks/AppKit.framework",
        "$SDKROOT/System/Library/Frameworks/Foundation.framework",
        "//VendorLib/SomePrebuilt:SomePrebuiltThing",
    ],

Reviewed By: williamtwilson

fbshipit-source-id: d91d113eeb