Optimize build and extension file parsing in Skylark parser.
Summary:
Previous implementation was doing traversal of an entire dependency tree
when discovering loaded paths, which was very inefficient especially for deep
dependency trees.

New implementation stores an entire transitive closure with extension itslef, so
that in order to compute transitive closure of any extension, only direct dependencies
have to be considered.

fbshipit-source-id: 23ff964647