Move (most) RE related config to RemoteExecutionConfig
Summary:
This moves all remote execution config not derived from the build
strategy to RemoteExecutionConfig. Changes that Config to a ConfigView,
and moves it to its own package.

The moved values are still configurable as `modern_build_rule.<key>`,
but will print an error if configured that way. I'll remove that
fallback after a short transition period.

Reviewed By: ttsugriy

fbshipit-source-id: 8948a848dd