Decouple RemoteExecution implementations from MBR strategies
Summary:
Currently the way to integrate remote execution into a build strategy is
by implementing RemoteExecution, which in turn implements
IsolatedExecution, both of which are tightly coupled to the specifics of
sending ModernBuildRule over remote execution. The implementations of
RemoteExecution though, just provide a remote execution service, a cas
service and a close() method.

This change introduces an interface RemoteExecutionClients in the
remoteexecution package that has those three functions (and a
getProtocol()). RemoteExecution is then changed from an abstract class
to a concrete class implemented in terms of a RemoteExecutionClients.
And the implementations are updated accordingly and moved to the
remoteexecution package.

Reviewed By: ruibm

fbshipit-source-id: f59b1f9713