Add flush_events_before_exit to ActionGraphProviderIntegrationTest
Summary:
ChromeTrace is written asynchronously when running buckd,
which meant that the file might not be materialized when we read
it for the test. It could also be the wrong version of the trace,
i.e. from the previous invocation instead of the current one.
This adds the buck-config to make chrome trace logging synchronous
so the test is correct

Reviewed By: ttsugriy

fbshipit-source-id: 25b0c6ce1c