Remove autodeps enable/disable option from BuckProjectSettings.
Summary:
The autodeps feature used to be a computationally expensive operation
that ran often on editor changes, so it was gated behind a project-level setting.
Now that it has been replaced with a quickfix (that must be accepted by the
user in order to trigger the computationally expensive part), there is no longer
a need to have a project setting for it.

Reviewed By: asp2insp

fbshipit-source-id: b06c7d2ce5