show all untracked header errors instead of only the first one
Summary:
# Issue
Buck can be set to force that every used header is listed on the *BUCK* file. When there are multiple unlisted/untracked headers only the first one is reported. This is inconvenient as one would like to get to know about all the untracked headers in one pass.

# Proposed solution
Accumulate the errors and throw the exception only after all headers have been checked.

# Output example:
Previously:
  two_untracked_headers.cpp: included an untracked header "untracked_header.h"

  Please reference these headers from "headers", "exported_headers" or "raw_headers"
  in the appropriate build rule.

With this diff:
  two_untracked_headers.cpp: included an untracked header "untracked_header.h"
  two_untracked_headers.cpp: included an untracked header "untracked_header_2.h"
  two_untracked_headers.cpp: included an untracked header "untracked_header_3.h"
  two_untracked_headers.cpp: included an untracked header "untracked_header_4.h"

  Untracked header/s detected. Please reference these headers
  from "headers", "exported_headers" or "raw_headers"
  in the appropriate build rule.

Reviewed By: cjhopman

fbshipit-source-id: 20b0c4e325