Use Map to store nodes processed from a build file in BuildFileRawNodeParsePipeline
Summary:
`BuildFileRawNodeParsePipeline` produces a list of raw nodes for every build file. `BuildTargetRawNodeParsePipeline` scans those lists to find targets using a linear search.

This change modifies the parsing pipeline to use a map with build target short names as keys to allow faster access to raw nodes.

For example, the following list of nodes:

```
[
  {
    "name": "a",
    ...
  },
  {
    "name": "b",
    ...
  }
]
```

It will be represented in the following form:

```
{
  "a" => {
    "name": "a",
    ...
  },
  "b" => {
    "name": "b",
    ...
  },
}
```

This form requires more memory, but provides faster access to individual targets. Another approach would be to use both linear search and map-based data structure and choose the approach based on the number of nodes in a build file; this would require maintaining two different types of daemon parser states though.

Reviewed By: ttsugriy

fbshipit-source-id: 51d1cfe6a5