Fix bug and improve include chain algorithm
Summary:
# Bug report
There's a bug that may trigger an **OOM** in the presence of include cycles in the project.

# Reason
When there's an untracked header *Buck* tries to generate an include chain for this header. While generating it the algorithm gets stuck in the cycle, effectively trying to construct an infinite chain.

# Proposed solution
We take advantage of the depfile format which allows to avoid cycles and simplify the algorithm.

Reviewed By: ilya-klyuchnikov

fbshipit-source-id: 295a38f0b7