Do not share executor service between pipelines to avoid deadlocks
Summary:
There are two pipelines to produce `TargetNode` - with and without configurable attributes. These pipelines share executor service, but the pipeline that supports configurable attributes can use blocking calls to get data from the other pipeline. In cases when the executor service has all the threads busy it leads to a deadlock.

A temporary solution is to use a separate executor service. In this case it's direct executor service which is fine given that all configuration rules are relatively fast to construct.

Reviewed By: ttsugriy

fbshipit-source-id: f793d34fb8