Add helper classes BuckTarget and BuckTargetPattern
Summary:
To enable consistent treatment of targets and target patterns in ideabuck, create utility classes to consistently parse and query targets and patterns.

Note that this duplication is necessary because we can't use the ones from buck, since those are (currently) entwined with buck's model/resolvers/etc., and ideabuck needs to be able to parse/reason about targets even when the buck graph isn't materialized (or even valid).

Reviewed By: asp2insp

fbshipit-source-id: 357336ea7e