Fix DefaultFileSystemProvider class loading issue on startup
Summary:
Certain Java10 builds of Buck crash with:
```
java.lang.ClassNotFoundException: com.facebook.buck.cli.bootstrapper.filesystem.BuckFileSystemProvider
```
This happens before Buck's `ClassLoaderBootstrapper` is loaded. Interestingly (and frustratingly), this doesn't repro if you try to attach a debugger with `-agentlib:jdwp`, which changes the the sequence of events during class loading.

Though we put the Buck bootstrapper JAR, which contains both `ClassLoaderBootstrapper` and `BuckFileSystemProvider`, on the regular classpath, the system classloader can't find this class when newer version of Java try to initialize `DefaultFileSystemProvider`, so we add the JAR to the boot class loader classpath.

Reviewed By: sbalabanov

fbshipit-source-id: 9e9890efbf