config: omit consecutive split chars when splitting config setting to a list
Summary:
A current use case is the use `$(config ...)` to embed one config setting inside
another in `.buckconfig`.  When used config settings meant to be lists, if the
referenced config setting is empty, this can cause an extra empty arg to be
parsed from the `.buckconfig`.  Since we already have support for explicitly
specifying an empty string using quotes, and since this is almost certainly not
the intended behavior, make sure we ignore consecutive split chars.

Reviewed By: philipjameson

fbshipit-source-id: 362eb1ec06