Improve the error message when encountering an unknown cell
Summary:
Admittedly the current error message is pretty good, but it has one glaring flaw:

_It doesn't tell you what target was using the invalid cell_

This is fine if the target names go from BUCK file to Buck directly, but if they go through any macros (especially macros that mess around with the cell name) then it becomes very difficult to debug issues. This diff is aimed at solving that problem, by adding a new type of exception and then catching that exception in the BuildTargetParser and providing a better error message.

Reviewed By: ttsugriy

fbshipit-source-id: 74e5853353