Allow to exclude targets by constraints when processing multiple targets
Summary:
When building targets using a pattern to specify multiple targets all the target will be built even though some targets may not be compatible with the target platform.

This change adds support for new attribute `target_compatible_with` which can be used to list constraints a target platform needs to match in order to build that target. This logic is enable only when a CLI option (`--exclude-incompatible-targets`) is passed.

Internally this change add `ParserWithConfigurableAttributes` which is a subclass of `DefaultParser` with some knowledge about options related to configurable attributes only.

Reviewed By: bobyangyf

fbshipit-source-id: d8953bb183