Give BuildRuleStrategies more power.
Summary:
Makes BuildRuleStrategy completely responsible for building the rule.
Previously this would delegate to BuildExecutorRunner.runWithExecutor().
That function took a BuildExecutor and delegated back to it. This was
used to run a small bit of code both before and after the BuildExecutor.

This pattern required the strategies to be implemented in such a way
that that would work correctly. Specifically, (1) they effectively
needed to be synchronous (or to sneakily do work outside of the
BuildExecutor) and (2) it's unclear how cancellation should work.

Recent changes have greatly decreased the amount of code that was being
run before and after.

This removes BuildExecutorRunner.runWithExecutor() and the BuildExecutor
class entirely. BuildExecutorRunner is renamed to BuildStrategyContext
and exposes the small pieces of information that are required to
implement runWithExecutor within the build strategies themselves.

Reviewed By: styurin

fbshipit-source-id: de6c0a4dcd