Fix RES_FOLDERS_RELATIVE_PATH to be actually relative (#2072)
Summary:
This changes the behavior of buck project to correctly set the `RES_FOLDERS_RELATIVE_PATH` value to be relative to the module. The android intellij plugin expects a proper relative path and does not work when macros like `$MODULE_DIR` etc. are prepended to the path.

This change is required to support Android Layout Preview in buck project and intellij.
Pull Request resolved: https://github.com/facebook/buck/pull/2072

Reviewed By: christolliday

Pulled By: christolliday

fbshipit-source-id: e4f8f76e14