Change dummy r dot java extra classpath to point to jars (#2071)
Summary:
Currently, buck project points the dummy r dot java output in the iml to the scratch directory. This is not always available especially when the outputs are materialized from the network cache. This changes the behavior to point to the final jar instead which will make sure the R classes can be found in all cases (from network, local build with both variations of `jar_spool_mode`).

This change is required to support Android Layout Preview in buck project and intellij.
Pull Request resolved: https://github.com/facebook/buck/pull/2071

Reviewed By: christolliday

Pulled By: christolliday

fbshipit-source-id: 478d96510c