Preserve the order of files in zip_file.zip_srcs
Summary:
The order of files in `zip_srcs` is important and should be preserved as it can affect which entry appears in the result zip file.

By changing the order of files in `zip_srcs` and adjusting the `on_duplicate_entry` parameter a user can control which entries will be copied to the result archive when some input archives contain entries with the same name.

Reviewed By: philipjameson

fbshipit-source-id: 571ba97e13