Fix buck with version resolution with configured buck-out directory
Summary:
Currently if you have multiple output directories configured, we only look at one main .currentversion to determine buck-out is cleaned up. This makes it so that if you change the buck-out directory, we make cleaning it contingent on that specific buck-out dir.

Original problem was something like this (with mode/opt pointing to an buck-out/opt, and mode/dev to buck-out/dev):
- buck build mode/opt //:foo
- buck build mode/dev //:foo
-- Breaking change in buck upgrade that changes files in buck-out, but not the rule keys
- buck build mode/dev //:foo # This cleans up buck-out/dev
- buck build mode/opt //:foo # This does /not/ clean up buck-out/opt because buck-out/.currentversion is already at the right version

Reviewed By: bobyangyf

fbshipit-source-id: a05f80b6a6