Add support for selectable attributes in buck query
Summary:
This change adds resolution of configurable attributes when `query` and `targets` command request JSON representation of target nodes.

This is implemented using the same approach as in resolution of configurable attributes during regular parsing, but the set of supported types is limited to those emitted by build file parsers (which includes int, string, map and list).

Reviewed By: bobyangyf

fbshipit-source-id: 3b6a3b1e32