Prevent hopeless fetches by bundle checking if artifact is present in the cache
Summary: If cache hit rate is low, we will spend a lot of time on trying to fetch each of the artifact (and failing). By bundling call to check if artifact are present, we can prevent "hopeless" fetches, and only process those which are in the cache. Moreover by bundling checks, we won't be limited by fetch concurrency rate.

Reviewed By: cjhopman

fbshipit-source-id: 9740e61215