Add additional_coverage_targets property to python_test
Summary: We would like to take a new attribute `additional_coverage_targets` into the PythonTest and use Buck to resolve the output path and annotate it in `external_runner_specs.json`. This piece of information will be later used by testpilot to do test coverage profile analysis. This diff works on this.

Reviewed By: styurin

fbshipit-source-id: d4a933c0e1