Added handling for Windows relativizing paths on different drives before calling Watchman.
Summary:
In cases where there are subst or ReparsePoints (RPs) on Windows, a path could be accessed through a RP that is really a path within a root on a different drive.
In such cases, relativization will fail because the WinFS is throwing when trying to relativize paths on different drives. Ask Watchman to resolve the final path and relativize it properly.

Reviewed By: bobyangyf

fbshipit-source-id: ece1d74184