Fix exit code to be BUILD_ERROR for stampede local builds
Summary:
Stampede local executors intercept all exceptions and generate exit codes. They should not be doing that because the appropriate error handling logic in Main is supposed to properly handle it, however it is a bigger refactoring to push exceptions up the stack.
This quick fix will make sure that for SOME user-errors a proper exit code is emitted.

Reviewed By: ruibm

fbshipit-source-id: f849e4dbcb