Add --show-output to audit actiongraph
Summary:
A number of action-graph rules will not show their output properly when
queried with `buck targets --show-output`. This diff allows augmenting the `audit actiongraph`
command with output paths.

This will aid manual introspection as well as allowing external tools such as linters or infer to
collect the classpath of a buck rule.

Reviewed By: sbalabanov

fbshipit-source-id: aa98ad9949