Add class to find includes for packages and subpackages
Summary:
Creates a finder that, given a path, finds the deepest path that has an implicit include. This is part of the work at allow a pre-configured .bzl file to be loaded for a given package and all subpackages.

e.g. If the root dir, and foo/bar are configured, baz/BUCK will return the root config, and foo/bar/BUCK and foo/bar/baz/BUCK both return the include at foo/bar

Reviewed By: styurin, ttsugriy

fbshipit-source-id: 04ddb08344