Add implicit_package_symbol method to skylark
Summary:
Adds a method that allows one to look up symbols that have been included for the specific package/subpackage in the configuration.

That is, when foo/BUCK is being parsed and is configured to implicitly include some symbols, and it uses //:something.bzl, something.bzl is able to access those symbols via implicit_package_symbol().

PythonDSL and docs will come in subsequent diffs.

Reviewed By: styurin, ttsugriy

fbshipit-source-id: 6b6daf2245