Fix BuckAnnotator for diff views
Summary:
The BuckAnnotator was throwing exceptions when annotating targets
in diff views because the backing file for the diff view was not
in the normal filesystem, but in a lightweight filesystem that
IntelliJ uses for holding temporary files during a diff process.
This diff does not fix the rendering in this view, but it at
least stops it from throwing exceptions.

Reviewed By: asp2insp

fbshipit-source-id: 8ebb202ecd