Delay path existense verification till the creation of target node
Summary:
Paths in rule attributes are validated for existence during coercion to `Path` in `PathTypeCoercer`. This happens for all attributes in a node including for configurable attributes. As a result, a path may not be used for a specific configuration, but its existence is verified during parse time. If that file doesn't exist, the build fails. At the same time users may use sparse profiles which do not check-in specific files, but since the build don't need excluded files they expect it to work.

The solution is to avoid checking for existence of a path in `PathTypeCoercer`, but check all the files references in a `TargetNode` before its creation in `TargetNodeFactory`.

Note that this change strictly speaking is more strict in situations when a description implements `ImplicitInputsInferringDescription` (which is implemented by `export_file` only). This interface allows descriptions to infer paths from other non-path attributes and, as a result, the number of checked files can be larger after this change.

To avoid rechecking the same paths multiple times it uses a cache of checked paths per project file system.

Reviewed By: bobyangyf

fbshipit-source-id: 3f6c74e118