BuildFileManifest and parsers to use ordered set for includes
Summary:
SkylarkProjectParserFile::getIncludes() has a bug - it returns full transitive set of all includes without deduplicating them, that causes the list to explode for deep dependency trees.
This change fixes that by changing the interface of includes to be ImmutableSortedSet.
This also improves strong fingerprint calculation for parsing cache runtime because it does not need to do expensive sorting of included files anymore.

fbshipit-source-id: f4633367e0