Be more defensive in CachingBuildRuleBuilder
Summary:
Adds a flag indicating whether or not we've called onOutputsWillChange.
This is verified as we finish the rule so that we require that it is set
if the success type indicates that the outputs have changed.

Also make BuildRuleSteps not implement RunnableWithFuture so it is
easier to see exactly where it's methods are being called.

Reviewed By: bobyangyf

fbshipit-source-id: 798caa6858