Move around where we do some work to prepare for running a rule (#2)
Summary:
Moves this setup out of executeCommandsNowThatDepsAreBuilt. It's only
important that this work happens before building a rule, it's not
important that it is immediately before. This is going to allow us to
let strategies completely handle how to build a rule without needing to
do this work.

Fixes issue with previous version.

Reviewed By: bobyangyf

fbshipit-source-id: a4020a5a2f