Make CachingBuildRuleBuilder cancellation explicit.
Summary:
This changes CachingBuildEngine to explicit notify all present
CachingBuildRuleBuilders of cancellation when the first error occurs (or
when cancelling for other explicit reasons).

This then changes BuildRuleStrategy to return a simple object that has
an explicit cancel() method in addition to the old Future that allows
CachingBuildRuleBuilder to forward those cancellation requests to the
strategy so that it can cancel pending work if it wants to.

Reviewed By: bobyangyf

fbshipit-source-id: 002c92e51c