Refactor+rename IsolatedExecutionStrategy and RemoteExecution
Summary:
Currently the construction of the information needed for remote
execution (the action, the inputs merkle tree) is oddly split across
these two classes. This was done so that the remote execution approach
could share logic with InProcessIsolatedExecution. Now that that has
been deleted, we can clean this up.

This change takes the code of those two classes, splits out everything
related to computing the remote execution action into
ModernBuildRuleRemoteExecutionHelper and all the logic for using remote
execution to implement a BuildRuleStrategy goes in
RemoteExecutionStrategy.

Reviewed By: bobyangyf

fbshipit-source-id: 06ba4feb6e