Remove grpc-netty and netty libraries
Summary:
We now have grpc-netty-shaded and it's odd (broken?) to use both
grpc-netty and grpc-netty-shaded. Also, the version of netty that we had
wasn't even the right version of netty for the version of grpc-netty
that we had :/

This removes grpc-netty and the standalone netty and just uses
grpc-netty-shaded everywhere that those were used.

fbshipit-source-id: 3034c030ca