Introduce hybrid_local build strategy
Summary:
This build strategy is pretty straightforward right now. It accepts 3
pieces of configuration `local_jobs`, `delegate_jobs`, and `delegate`.
`delegate` configures another build strategy to delegate overflows to.
As build requests come in, it either builds them locally (with the
default build behavior), sends them to the delegate strategy, or puts
them in a queue for either the currently active local jobs or the
delegate jobs to fall below the configured thresholds.

This approach serves two purposes:

1. It allows the build to use local resources w/ remote execution
enabled
2. It may avoid remote execution overhead on the critical path

Reviewed By: bobyangyf

fbshipit-source-id: 64b88c179e