Improve error message when parsing target pattern with unknown cell
Summary: A month or so ago I changed the `BuildTargetParser` so that it would tell you what target had the missing cell, if it couldn't find one. This has been useful, but I recently ran into the same problem. This time the caller was the `BuildTargetPatternParser`, so give it the same treatment here.

Reviewed By: styurin

fbshipit-source-id: d6e022639d