DependencyAggregation: export deps as runtime deps
Summary:
We use the `DependencyAggregation` rule a optimization to avoid attaching same
the large number of preprocessor deps to each compilation rule of a given library.
However, there's a bug with the current setup, as if the aggregation rule has a
matching rule key hit, it won't verify that it's deps are also up-to-date and
so may leave them in a stale state when running the compilation.

This diff exports the `DependencyAggregation` deps as runtime deps, so that they
are always processed.

Reviewed By: cjhopman

fbshipit-source-id: 709dfc03c6