Fix NullPointerException in Jasabi with Java 10
Summary: In Java 10+, the compiler APIs will iterate over static initializers for classes in cases where they wouldn't before. It's valid for `ClassVisitor::visitMethod` to return `null`, and Buck's implementation in `AbiFilteringClassVisitor` does just this for static initializers, as we don't care about them for the purposes of ABI generation. Add a null check in the caller.

Reviewed By: cjhopman

fbshipit-source-id: d2a2ebb0a7