Implement HybridLocal work stealing
Summary:
Currently there's a big tradeoff if we want to send a lot of work to the
delegate. Once we send it there, we can't run it locally even if local
resources become free after we send it. For remote execution, especially
as we start to send 100s of jobs to the strategy, there may be a
significant period of time between when we send it to the strategy and
when it actually begins execution. During that period, it would be best
if we could move it back to running locally if resources become free.

This adds BuildRuleStrategyResult.cancelIfNotStarted which will cancel a
build task if it "hasn't started" (which is up to the strategy to
decide).

Using that, this adds job stealing to the HybridLocalStrategy. This is
pretty straightforward, when scheduling a job to the delegate we add it
to a queue of possible jobs to steal. When a scheduleLocal job is
started, if there's no jobs in the overflow queue, we'll try to steal
jobs from the queue of jobs sent to the delegate.

Reviewed By: bobyangyf

fbshipit-source-id: dac1d2aea8