Improve UI for Remote Execution actions
Summary:
Now that the remote execution strategy has multiple async phases, the UI
has some behavior that is hard to understand. For example, the ui might
disappear when all actions are currently in queues in-between phases.

This adds an initial "scheduled" event so that the listener can know how
many nodes have started. It then adds a "waiting" state that keeps track
of each node that has started but is not in one of the other active
states.

Reviewed By: ruibm

fbshipit-source-id: 8d69b8bb32