Change InputsVisitor to InputsMapBuilder
Summary:
This changes inputs calculation to cache computed nodes similar to
serialization does (and rk calculations, too).

The main thing that this will allow us to do is to derive merkle trees
for each rule key appendable and then merge those instead of recomputing
full merkle trees for every rule. On a decently sized graph, that
approach to merkle tree construction is as much as 30x faster.

Reviewed By: jeffdunn

fbshipit-source-id: feb7407296