Add MerkleTreeNodeCache
Summary:
This is used to create a tree data structure for a set of files/symlinks
such that nodes in the tree are interned. It provides functionality to
compute and cache the protocol-encoded data to use these merkle trees as
inputs (and outputs) for remote execution.

We'll use this to compute merkle trees for every rulekey appendable,
where each one is a merge of the tree for its directly referenced inputs
and the merkle trees of referenced appendables. Because merging
non/slightly overlapping trees is efficient (both space and time) this
will be great.

This switches some of the really simple uses of FileTreeBuilder over to
MerkleTreeNodeCache (but leaves the complicated one).

The actual builder logic here is quite similar to
FileTreeBuilder/ProtocolTreeBuilder but operating directly on
FileNode/SymlinkNode, using sorted maps, and handling interning. The
next diffs are going to switch all remaining uses of FileTreeBuilder to
MerkleTreeNodeCache.

Reviewed By: bobyangyf

fbshipit-source-id: 5aa5a129ed