Create action information much, much faster
Summary:
This changes ModernBuildRuleRemoteExecutionHelper to use
InputsMapBuilder directly to get the inputs-relevant subgraph of rulekey
appendables and then MerkleTreeNodeCache to construct merkle trees of
inputs in ways that takes advantage of the extensive sharing of rulekey
appendables between different rules. In a large app, this cuts
action creation time from 1.33s/action to 0.016s/action.

Now that FileTreeBuilder isn't used, it's deleted.

Reviewed By: bobyangyf

fbshipit-source-id: 6fed8b5ab0